+++ src/views/HomeView.vue
@@ -5,27 +5,7 @@
         <div class="sections_container">
             <About />
             <Portfolio />
-            <section class="services main_section">
-                <section class="text">
-                    <h3>Services</h3>
-                    <p>
-                        {{ services_description }}
-                    </p>
-                    <ul class="services_list">
-                        <li v-for="item in services">
-                            <div class="service-item">
-                                <i class="fa-solid fa-caret-right"></i>
-                                <h4>{{ item.title }}</h4>
-                            </div>
-                            <p>{{ item.description }}</p>
-                        </li>
-                    </ul>
-                </section>
-                <section class="photos">
-                    <img :src="service_photo_1" alt="service photo 1" />
-                    <img :src="service_photo_2" alt="service photo 2" />
-                </section>
-            </section>
+            <Services />
             <section class="contact main_section">
                 <div class="photo">
                     <img :src="contact_photo" alt="contact photo" />
@@ -61,41 +41,17 @@
 import { useViewStore } from '@/stores/views'
 
 // TODO: Update photos with some real ones from Julia's gallery
-import service_photo_1 from '@/assets/images/services_1.jpg'
-import service_photo_2 from '@/assets/images/services_2.jpg'
+
 import contact_photo from '@/assets/images/contact.jpg'
 
 import SocialMedia from '@/components/UI/SocialMedia.vue'
 import ContactForm from '../components/ContactForm.vue'
-import Portfolio from '@/components/Portfolio.vue'
 import About from '@/components/About.vue'
+import Portfolio from '@/components/Portfolio.vue'
+import Services from '@/components/Services.vue'
 
 const viewStore = useViewStore()
 const views = viewStore.views
-
-const services_description =
-    'Looking for a videographer for your next project? I offer a range of services, including:'
-
-const services = [
-    {
-        title: 'Short Film Production',
-        description: 'from concept to post-production.',
-    },
-    {
-        title: 'Promotional Videos',
-        description:
-            'creation of promotional videos for products like: cosmetics, clothes, books, and all you can think about!',
-    },
-    {
-        title: 'Video Editing',
-        description:
-            'if you have existing footage, I can edit it into a polished, professional video.',
-    },
-    {
-        title: 'Scenario writing',
-        description: 'scenarios for short films 2-3 min. ',
-    },
-]
 
 const email = 'iu.matiash@gmail.com'
 </script>
@@ -114,64 +70,6 @@ const email = 'iu.matiash@gmail.com'
 
         .main_section {
             margin: 1rem 2rem;
-        }
-
-        .services {
-            margin-top: 5rem;
-            margin-bottom: 5rem;
-            display: flex;
-            text-align: justify;
-
-            .text {
-                ul {
-                    list-style: none;
-                    margin-top: 2rem;
-                    margin-right: 2rem;
-                    padding: 0;
-
-                    li {
-                        margin-right: 1rem;
-                        margin-top: 2rem;
-                        align-items: center;
-
-                        .service-item {
-                            display: flex;
-                            align-items: center;
-                            margin-right: 1rem;
-                            margin-top: 2rem;
-                            margin-bottom: 0.5rem;
-
-                            i {
-                                font-size: 1.5rem;
-                                margin-right: 1rem;
-                            }
-
-                            h4 {
-                                margin: 0;
-                            }
-
-                            p {
-                                font-size: 1.2rem;
-                            }
-                        }
-                    }
-                }
-            }
-
-            .photos {
-                width: 50%;
-                margin-top: 2rem;
-                display: flex;
-                flex-wrap: wrap;
-                justify-content: space-between;
-
-                img {
-                    width: 48%;
-                    height: auto;
-                    border-radius: 5%;
-                    object-fit: cover;
-                }
-            }
         }
 
         .contact {
@@ -245,34 +143,6 @@ const email = 'iu.matiash@gmail.com'
         margin: 0;
 
         .sections_container {
-            .services {
-                margin-top: 2rem;
-                margin-bottom: 2rem;
-                flex-direction: column;
-
-                .text {
-                    ul {
-                        margin-right: 0;
-                    }
-                }
-
-                .photos {
-                    width: 100%;
-                    margin-top: 2rem;
-                    margin-bottom: 2rem;
-                    display: flex;
-                    flex-wrap: wrap;
-                    justify-content: space-between;
-
-                    img {
-                        width: 45%;
-                        height: auto;
-                        border-radius: 5%;
-                        object-fit: cover;
-                    }
-                }
-            }
-
             .contact {
                 padding-top: 2rem;
                 padding-bottom: 2rem;
